Rename 'embedly' to 'proxy' #143
