+++ .coveragerc
@@ -1,1 +1,1 @@
-embedly/.coveragerc
+app/.coveragerc
