+++ app/.coveragerc
@@ -1,7 +1,7 @@
 # .coveragerc to control coverage.py
 [run]
 include =
-    embedly/*
+    proxy/*
 
 [report]
 # Regexes for lines to exclude from consideration
