--- app/version.json
@@ -1,1 +1,0 @@
-{"commit":"3b9a7a5eb1e916f471ad13fa5200b23441b16cf3","version":"0.12-13-g3b9a7a5","source":"https://github.com/sarracini/embedly-proxy.git"}
